requirements.txt
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.6
anaconda-project==0.8.3
appdirs==1.4.3
argh==0.26.2
asgiref==3.2.7
asn1crypto==1.3.0
astroid==2.4.1
astropy==4.0.1.post1
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.4
blinker==1.4
bokeh==2.0.2
boto==2.49.0
boto3==1.12.21
botocore==1.15.21
Bottleneck==1.3.2
certifi==2019.3.9
cffi==1.14.0
chardet==3.0.4
ChatterBot==0.8.7
chatterbot-corpus==1.1.4
click==7.1.2
cloudpickle==1.4.1
clyent==1.2.2
colorama==0.4.3
comtypes==1.1.7
conda==4.8.3
conda-build==3.17.6
conda-package-handling==1.7.0
conda-verify==3.1.1
contextlib2==0.6.0.post1
couch==1.0.1
CouchDB==1.2
couchquery==0.10.3
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.17
cytoolz==0.10.1
dask==2.17.2
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
distributed==2.17.0
dj-database-url==0.5.0
Django==2.2.14
django-allauth==0.42.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-dynamodb-sessions==0.6
django-heroku==0.3.1
django-phone-field==1.8.1
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
flake8==3.8.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-paginate==0.6.0
Flask-WTF==0.14.3
fsspec==0.7.4
future==0.18.2
gevent==1.4.0
gevent-websocket==0.10.1
glob2==0.7
greenlet==0.4.15
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
h5py==2.8.0
HeapDict==1.0.1
html5lib==1.0.1
httplib2==0.17.3
hypothesis==5.11.0
idna==2.9
imageio==2.8.0
imagesize==1.2.0
importlib-metadata==1.6.0
imutils==0.5.1
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.15.1
json5==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.1.4
keyring==21.1.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.32.1
locket==0.2.0
lxml==4.5.0
MarkupSafe==1.1.1
mathparse==0.1.2
matplotlib==3.1.3
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.3.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.49.1
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
oauthlib==2.1.0
olefile==0.46
openpyxl==3.0.3
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
partd==1.1.0
path==13.2.0
path.py==12.0.2
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
phonenumbers==8.12.6
pickleshare==0.7.5
Pillow==7.1.2
pipenv==2018.11.26
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psutil==5.7.0
psycopg2==2.8.5
py==1.8.1
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.5
pydocstyle==4.0.1
pyflakes==2.2.0
pygame==1.9.4
Pygments==2.6.1
pylint==2.5.2
pymemcache==3.1.0
pymongo==3.7.2
PyNaCl==1.3.0
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyotp==2.3.0
pyparsing==2.4.7
PyQt5-sip==12.7.2
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.2
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.5.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
python-twitter==3.5
python3-memcached==1.51
python3-openid==3.2.0
pytz==2018.5
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8.1
QtAwesome==0.7.0
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.1.0
rope==0.17.0
Rtree==0.9.4
ruamel-yaml==0.15.87
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.1
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==2.0.1
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.1
spyder==4.1.3
spyder-kernels==1.9.1
SQLAlchemy==1.3.17
sqlparse==0.2.4
statsmodels==0.11.1
sympy==1.5.1
tables==3.4.4
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.8
virtualenv==20.0.20
virtualenv-clone==0.5.4
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
WTForms==2.2.1
xlrd==1.2.0
XlsxWriter==1.2.8
xlwings==0.19.4
xlwt==1.3.0
yapf==0.28.0
zict==2.0.0
zipp==3.1.0
